Use a sqlite database to store channels
Using a sqlite DB enables sharing SQL schema and actual datas with the scala
implementation.

The SQLDelight library is used to generate type safe kotlin methods from
sql statements and schemas. A `ChannelsDatabase.sq` file is created to
describe the database.

Note that the channels database is a physical file. On Android this file
will be stored in the app's private directory > databases folder.