The client stores and provides the electrum config
The shared module should not handle the storage and selection of the
user preferences. This is done by the app that can leverage the platform
standard preferences utilities.

The electrum controller has been simplified a bit, as well as the
configuration manager. The wallet xpub is now computed from the
Wallet object.

The kodein-db dependency has been removed.