Add incoming balance to home controller
The home model now exposes the balance that is incoming. It will hold
the balance for the pending swap-ins (useful because of the delay for
swaps). When a swap-in confirms on that address, the incoming balance is
updated.

This commit also simplifies the logic updating the payments in the
payment manager: instead of listening to payment events from the
lightning engine, we watch the database using a sqldelight lib feature
that publish a flow and update it when the data change. The home controller
in turn collects this flow to populate the payments list in its model.