Handle pay-to-open payments where the amount contain the fee
This commit depends on https://github.com/ACINQ/lightning-kmp/pull/290

Old pay-to-open (L2 -> L2 requiring a new channel to be created) stored
the fee in the amount field which is not correct anymore (see PR linked
above). To fix that this commit adds a new MULTIPART_V1 version and uses
the receivedWith deserializer to discriminate the old pay-to-open payments
from the rest and subtract the fee from the amount in that case.

Fixes https://github.com/ACINQ/phoenix/issues/170